+++ src/Ampersand/Commands/Proto.hs
@@ -14,20 +14,10 @@ import           Ampersand.FSpec
 import           Ampersand.Misc.HasClasses
 import           Ampersand.Prototype.GenFrontend
 import           Ampersand.Types.Config
-import qualified RIO.Text as T
 import           System.Directory
 -- | Builds a prototype of the current project.
-proto :: ( Show env
-         , HasRunner env
-         , HasFSpecGenOpts env
-         , HasDirCustomizations env
-         , HasZwolleVersion env
-         , HasProtoOpts env
-         , HasDirPrototype env
-         , HasGenerateFrontend env
-         , HasGenerateBackend env
-         , HasGenerateMetamodel env
-         ) 
+--
+proto :: (Show env, HasRunner env, HasFSpecGenOpts env, HasProtoOpts env, HasDirPrototype env, HasGenerateFrontend env, HasGenerateBackend env, HasCheckCompilerVersion env, HasGenerateMetamodel env)
        => FSpec -> RIO env ()
 proto fSpec = do
     env <- ask
@@ -43,9 +33,6 @@ proto fSpec = do
       then do doGenBackend fSpec
       else do logDebug "  Skipping generating backend files"
     generateMetamodel <- view generateMetamodelL
-    if generateMetamodel
+    if generateMetamodel --TODO @stefjoosten, Why should this be generated at the proto command??
       then do doGenMetaModel fSpec
       else do logDebug "  Skipping generating metamodel.adl"
-    copyCustomizations
-    dirPrototypeA <- liftIO $ makeAbsolute dirPrototype
-    logInfo $ "Prototype files have been written to " <> display (T.pack dirPrototypeA)